+++ reservation/src/main/java/table/eat/now/reservation/reservation/infrastructure/messaging/kafka/listener/dto/EventType.java
@@ -2,7 +2,7 @@
  * @author : jieun(je-pa)
  * @Date : 2025. 04. 20.
  */
-package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listner.dto;
+package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listener.dto;
 
 public enum EventType {
   RESERVATION_PAYMENT_SUCCEED,
