+++ reservation/src/main/java/table/eat/now/reservation/reservation/infrastructure/messaging/kafka/listener/dto/KafkaCommonEvent.java
@@ -1,0 +1,7 @@
+package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listener.dto;
+
+public interface KafkaCommonEvent {
+
+  EventType eventType();
+
+}
