+++ reservation/src/main/java/table/eat/now/reservation/reservation/infrastructure/messaging/kafka/listener/dto/PaymentEvent.java
@@ -1,0 +1,5 @@
+package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listener.dto;
+
+public interface PaymentEvent {
+  String paymentUuid();
+}
