--- reservation/src/main/java/table/eat/now/reservation/reservation/infrastructure/messaging/kafka/listner/dto/KafkaCommonEvent.java
@@ -1,7 +1,0 @@
-package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listner.dto;
-
-public interface KafkaCommonEvent {
-
-  EventType eventType();
-
-}
