--- reservation/src/main/java/table/eat/now/reservation/reservation/infrastructure/messaging/kafka/listner/dto/PaymentEvent.java
@@ -1,5 +1,0 @@
-package table.eat.now.reservation.reservation.infrastructure.messaging.kafka.listner.dto;
-
-public interface PaymentEvent {
-  String paymentUuid();
-}
